added noise to toy dataset

+++ lib/toy_data.py
@@ -83,11 +83,12 @@ def inf_train_gen(data, rng=None, batch_size=200):
     elif data == "rowimg":
         imagewidth = 200
         blockwidth = 20
-        x = np.zeros([batch_size,imagewidth])
+        x = np.zeros([batch_size,imagewidth],dtype=np.float64)
         for sample in x:
             o1 = np.random.randint(0,imagewidth-blockwidth)
             o2 = np.random.randint(0,imagewidth-blockwidth)
-            sample[o1:o1+blockwidth] +=1
-            sample[o2:o2+blockwidth] +=1
+            sample[o1:o1+blockwidth] +=1.
+            sample[o2:o2+blockwidth] +=1.
+
         x += np.random.randn(*x.shape) * 0.1
         return x
